Move AttachFileControl to its own file
